CineVerse-Backend/requirements.txt
fastapi
uvicorn
pydantic
email-validator
pandas
scipy
scikit-learn
openai
httpx
python-dotenv
rapidfuzz
pandas